Move inner accordion component to own file
