+++ context/app/static/js/components/Search/filters/FilterInnerAccordion/FilterInnerAccordion.jsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import Typography from '@material-ui/core/Typography';
+
+import { InnerAccordionDetails, InnerAccordionSummary } from './style';
+import { StyledAccordion, StyledExpandMoreIcon } from '../../styleForSidebar';
+
+function InnerAccordionContainer({ title, children }) {
+  if (Array.isArray(children) && children[0].props.items.length === 0) {
+    // Presumably a list filter...
+    return null;
+  }
+  if (
+    !Array.isArray(children) &&
+    children.props.items?.map((bucket) => bucket.doc_count).every((count) => count === 0)
+  ) {
+    // Presumably a range filter...
+    return null;
+  }
+  return (
+    <StyledAccordion key={title} defaultExpanded>
+      <InnerAccordionSummary expandIcon={<StyledExpandMoreIcon />}>
+        <Typography variant="subtitle2" color="textPrimary">
+          {title}
+        </Typography>
+      </InnerAccordionSummary>
+      <InnerAccordionDetails id={title.replace(/\s/g, '-')}>{children}</InnerAccordionDetails>
+    </StyledAccordion>
+  );
+}
+
+export default InnerAccordionContainer;
