+++ context/app/static/js/components/Search/filters/FilterInnerAccordion/index.js
@@ -1,0 +1,3 @@
+import FilterInnerAccordion from './FilterInnerAccordion';
+
+export default FilterInnerAccordion;
